+++ packages/core/src/index.ts
@@ -7,8 +7,20 @@ import Interweave, { InterweaveProps } from './Interweave';
 import Markup, { MarkupProps } from './Markup';
 import Filter, { FilterInterface } from './Filter';
 import Matcher, { MatcherInterface } from './Matcher';
+import Element from './Element';
+import Parser from './Parser';
 
-export { InterweaveProps, Markup, MarkupProps, Filter, FilterInterface, Matcher, MatcherInterface };
+export {
+  InterweaveProps,
+  Markup,
+  MarkupProps,
+  Filter,
+  FilterInterface,
+  Matcher,
+  MatcherInterface,
+  Element,
+  Parser,
+};
 
 export * from './types';