+++ packages/core/tests/Matcher.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Element from '../src/Element';
-import { CodeTagMatcher, MockMatcher } from '../../../tests/mocks';
+import { CodeTagMatcher, MockMatcher } from '../src/testUtils';
 
 describe('Matcher', () => {
   const matcher = new CodeTagMatcher('foo', '1');